+++ v3/aevo/src/Hasher/contracts/interfaces/IHasher.sol
@@ -1,10 +1,6 @@
 // SPDX-License-Identifier: GPL-3.0-only
 pragma solidity 0.8.7;
 
-/**
- * @title IHasher
- * @notice Interface for hasher contract that calculates the packed message
- */
 interface IHasher {
     /**
      * @notice returns the bytes32 hash of the message packed
@@ -18,9 +14,9 @@ interface IHasher {
      * @param payload the data packed which is used by inbound for execution
      */
     function packMessage(
-        uint32 srcChainSlug,
+        uint256 srcChainSlug,
         address srcPlug,
-        uint32 dstChainSlug,
+        uint256 dstChainSlug,
         address dstPlug,
         bytes32 msgId,
         uint256 msgGasLimit,
