+++ v3/aevo/src/Hasher/contracts/interfaces/IPlug.sol
@@ -1,10 +1,6 @@
-// SPDX-License-Identifier: GPL-3.0-only
+// SPDX-License-Identifier: UNLICENSED
 pragma solidity 0.8.7;
 
-/**
- * @title IPlug
- * @notice Interface for a plug contract that executes the message received from a source chain.
- */
 interface IPlug {
     /**
      * @notice executes the message received from source chain
@@ -13,7 +9,7 @@ interface IPlug {
      * @param payload_ the data which is needed by plug at inbound call on remote
      */
     function inbound(
-        uint32 srcChainSlug_,
+        uint256 srcChainSlug_,
         bytes calldata payload_
     ) external payable;
 }
