+++ v3/dolomite/src/SimpleIsolationModeWrapperTraderV2-2/contracts/protocol/interfaces/IDolomiteStructs.sol
@@ -171,7 +171,7 @@ interface IDolomiteStructs {
 
     struct Par {
         bool sign;
-        uint128 value;
+        uint256 value;
     }
 
     struct RiskLimits {
